+++ addOns/spiderAjax/CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Update minimum ZAP version to 2.16.0.
 - Updated automation framework documentation and templates for `spiderAjax` job to reflect changes to the default value of numberOfBrowsers parameter
 - Fields with default or missing values are omitted for the `spiderAjax` job in saved Automation Framework plans.
+- Default the number of browsers to the number of available cores.
 
 ## [23.21.0] - 2024-09-02
 ### Changed